requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
django-appconf==1.0.5
django-configurations==2.4
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-solo==2.0.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
filelock==3.9.0
flake8==6.0.0
icalendar==5.0.4
identify==2.5.17
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
nodeenv==1.7.0
oauthlib==3.2.2
openpyxl==3.1.1
packaging==23.0
Pillow==9.4.0
platformdirs==3.0.0
pre-commit==3.0.4
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
XlsxWriter==3.0.8
mypy
flake8